Add full project file, fix leaderboard, add cmds
Here's the full project file (everything needed to build the bot), as well
as fixed some issue with the leaderboard calculation. There's a bug with
the same player guessing in multiple channels now, but it's worked around
so the bot doesn't crash.

I've also added !about and !help, as people were asking such questions.